backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
chromadb
python-multipart==0.0.6
ollama==0.1.7
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
html2text==2020.1.16